requirements.txt
PyMuPDF
Pillow
transformers
matplotlib
tabulate
torch
